PR from main

+++ Backend/Routes/Dependent/CourseLearningJourneyRoutes.py
@@ -14,7 +14,7 @@ from random import randint
 @app.post("/courselearningjourney/seedall", tags=["CourseLearningJourney", "SeedAll"])
 def seedAll():
     try:
-    # get learning journey
+        # get learning journey
         session = Session(engine)
         getLearningJourneyStatement = select(LearningJourneyModel)
         allLearningJourneys = session.exec(getLearningJourneyStatement).all()
@@ -44,6 +44,7 @@ def seedAll():
             "messaage": str(e)
         }
 
+
 @app.delete("/courselearningjourney/deleteall", tags=["CourseLearningJourney", "DeleteAll"])
 def deleteAll():
     return deleteAllData(CourseLearningJourneyModel)
@@ -65,7 +66,6 @@ async def AddCourseLearningJourney(request: Request, session: Session = Depends(
             learningJourney = chosenLearningJourney[0]
         if learningJourney == None:
             errors.append("learningJourney does not exist!")
-            errors.append(str(e))
             return {
                 "success": False,
                 "message": errors
@@ -85,9 +85,9 @@ async def AddCourseLearningJourney(request: Request, session: Session = Depends(
         session.commit()
         session.close()
         return {
-            "success": True,
-            "message": "Course(s) assigned to Learning Journey"
-        }
+                "success": True,
+                "message": "Course(s) assigned to Learning Journey"
+            }
 
     except Exception as e:
         errors.append(str(e))
@@ -96,31 +96,39 @@ async def AddCourseLearningJourney(request: Request, session: Session = Depends(
             "message": errors
         }
 
+
 @app.delete("/courselearningjourney/", tags=["CourseLearningJourney"])
 async def deleteCourseLearningJourney(request: Request, session: Session = Depends(get_session)):
     errors = []
     try:
         # find the course
         requestData = await request.json()
-        selectedCourseStatement = select(CourseModel).where(CourseModel.Course_ID == requestData["Course_ID"])
+        selectedCourseStatement = select(CourseModel).where(
+            CourseModel.Course_ID == requestData["Course_ID"])
         selectedCourse = session.exec(selectedCourseStatement).all()
-        if(len(selectedCourse) == 0):
-            errors.append("Couse does not exist!")
-        elif(len(selectedCourse) == 1):
+        if (len(selectedCourse) == 0):
+            errors.append("Course does not exist!")
+        elif (len(selectedCourse) == 1):
             errors.append("Only one course left cannot delete!")
         else:
             selectedCourse = selectedCourse[0]
 
-
         # delete all the relations
         # hard delete the relations
         allRelatedRelationStatement = select(CourseLearningJourneyModel).where(
-            CourseLearningJourneyModel.Course_ID == requestData["Course_ID"])
+            CourseLearningJourneyModel.Course_ID == requestData["Course_ID"] and CourseLearningJourneyModel.LearningJourney_ID == requestData["LearningJourney_ID"])
         relationResults = session.exec(allRelatedRelationStatement)
         allRelations = relationResults.all()
+        if len(allRelations) == 0:
+            errors.append("Course not inside learning journey!")
+        if len(errors) > 0:
+            return {
+                "success": False,
+                "message": errors
+            }
         for relation in allRelations:
             session.delete(relation)
-            
+
         session.commit()
         session.close()
         return {
@@ -128,7 +136,7 @@ async def deleteCourseLearningJourney(request: Request, session: Session = Depen
             "message": "Course Learning journey deleted"
         }
     except Exception as e:
-        errors.append(str(e))
+        errors.append(e)
         return {
             "success": False,
             "message": errors
